Storybook: ComposedChart - refactoring
