+++ storybook/README.md
@@ -1,0 +1,13 @@
+We are following the Component Story Format 3.
+Find more information about it here. https://storybook.js.org/blog/storybook-csf3-is-here
+
+If you are looking for a template from which to write stories, please do align with the ComposedChart.stories.tsx
+
+- We want to use templates for rendering, and multiple stories with different parameters, content, data.
+
+- We want to document every single component in
+-- the most simple manner
+-- a complex usage
+-- edge cases
+
+- Additionally we will want to add stories that document complex usage of recharts with custom components.
